+++ tests/test_package_layout.py
@@ -36,6 +36,7 @@ PUBLISHED_BROKERS = (
     'django_aiogram.broker.redis_list.RedisListBroker',
     'django_aiogram.broker.redis_streams.RedisStreamsBroker',
     'django_aiogram.broker.rabbitmq.RabbitMQBroker',
+    'django_aiogram.broker.kafka.KafkaBroker',
 )
 
 PUBLISHED = (